test(formatter): consolidate redundant TS fixtures (#26125)
Consolidate redundant fixtures.
