--- crates/oxc_formatter/tests/fixtures/ts/assignments/issue-17804.ts
@@ -1,2 +1,0 @@
-const fooRef =
-        useRef<Record<string, LazyFooThingFD<T, TError> | null | undefined>>(cache);
