--- crates/oxc_formatter/tests/fixtures/ts/assignments/issue-18927.ts
@@ -1,4 +1,0 @@
-class A {
-    readonly customHeaderTemplate =
-      viewChild.required<TemplateRef<{ total: number }>>("customHeader");
-}
