--- crates/oxc_formatter/tests/fixtures/ts/assignments/short-argument.ts
@@ -1,2 +1,0 @@
-const requestTrie =
-  TernarySearchTree.forPaths<IRecursiveWatchRequest>(!isLinux);
