--- crates/oxc_formatter/tests/fixtures/ts/directives/issue-21152-block-comment-multiline.ts
@@ -1,6 +1,0 @@
-"use client"; /* multi
-line
-comment */
-
-import { Foo } from "foo";
-import { Bar } from "bar";
