--- crates/oxc_formatter/tests/fixtures/ts/directives/issue-21152-block-comment.ts
@@ -1,4 +1,0 @@
-"use client"; /* single line block comment */
-
-import { Foo } from "foo";
-import { Bar } from "bar";
