--- crates/oxc_formatter/tests/fixtures/ts/directives/issue-21152.ts
@@ -1,4 +1,0 @@
-"use client"; // browser only
-
-import { Foo } from "foo";
-import { Bar } from "bar";
