+++ crates/oxc_formatter/tests/fixtures/ts/parameters/this.ts
