--- crates/oxc_formatter/tests/fixtures/ts/semi/options.json
@@ -1,8 +1,0 @@
-[
-  {
-    "semi": false
-  },
-  {
-    "semi": true
-  }
-]
