+++ crates/oxc_formatter/tests/fixtures/ts/semicolons/method-signatures.ts
