--- crates/oxc_formatter/tests/fixtures/ts/type-literal/params.ts
@@ -1,4 +1,0 @@
-export default function useTagsCount({
-  query,
-}: { query?: Record<any, any> } = {}) {
-}
