--- crates/oxc_formatter/tests/fixtures/ts/union/issue-20219.ts
@@ -1,32 +1,0 @@
-// Issue #20219 - non-idempotent formatting: leading JSDoc comment on single union member
-export type AuditLogOrderField =
-  /** Order audit log entries by timestamp */
-  | 'CREATED_AT';
-
-// Multiple JSDoc blocks before single union member
-export type MultipleJSDocBlocks =
-  /** Comment 1 */
-  /** Comment 2 */
-  /** Comment 3 */
-  | 'CREATED_AT';
-
-// Multiline JSDoc block before single union member
-export type MultilineJSDoc =
-  /**
-   * Order audit log entries by timestamp.
-   * This is a multiline comment.
-   */
-  | 'CREATED_AT';
-
-// Line comment before single union member
-type LineComment =
-  // line comment
-  | 'VALUE';
-
-// Intersection type with own-line comment
-type IntersectionComment =
-  /** JSDoc */
-  & 'VALUE';
-
-// Inline comment stays after `=` (not own-line, never moved before `=`)
-type InlineComment = /*1*/ | C;
