--- crates/oxc_formatter/tests/fixtures/ts/union/issue-21399.ts
@@ -1,15 +1,0 @@
-// Issue #21399 - comment after pipe in single member union
-export type myType = |
-  // Comment
-  "A";
-
-// Block comment after pipe on own line
-type BlockComment = |
-  /* block comment */
-  "B";
-
-// Multiple comments after pipe
-type MultipleComments = |
-  // Comment 1
-  // Comment 2
-  "C";
