--- crates/oxc_formatter/tests/fixtures/ts/union/parenthesis.ts
@@ -1,2 +1,0 @@
-type T1<B> = | (B extends any ? number : string);
-type T2 = | (() => void);
